requirement.txt
streamlit==1.54.0
fastapi==0.128.0
pydantic==2.12.5
uvicorn==0.40.0
mysql-connector-python==9.5.0
pandas==2.2.3
requests==2.32.3
pytest==9.0.2




